+++ src/extension.ts
@@ -1,6 +1,7 @@
 // IDEA(stefano): implement multiple copy/paste buffers
 // IDEA(stefano): implement visual line mode commands
-// IDEA(stefano): add padding to mode names to avoid shifting the rest of the icons in the bar everytime the mode changes
+// IDEA(stefano): implement cursor alignment, to remove the "Cursor Align" extension
+// IDEA(stefano): implement toggling of quote kinds, to remove the "Toggle Quotes"
 
 import * as vscode from "vscode";
 
@@ -10,21 +11,21 @@ declare global {
     }
 }
 
-interface ModeProperties {
-    readonly name: string;
-    readonly capturing: boolean;
-}
-
 class Mode {
     public name: string;
     public text: string;
 
     constructor(name: string) {
-        // TODO(stefano): trim the name to remove whitespace
         this.name = name;
         this.text = `-- ${name} --`;
     }
 }
+
+const SELECT_MODE_COMMAND = "modalcode.select_mode";
+const SELECT_MODE_TOOLTIP = "Select mode";
+const SELECT_MODE_PLACEHOLDER = "Select mode to enter";
+const ENTER_MODE_COMMAND = "modalcode.enter_mode";
+const MODE_CONTEXT_KEY = "modalcode.mode";
 
 let modes_names: string[];
 let modes: Mode[];
@@ -34,72 +35,81 @@ let status_bar_item: vscode.StatusBarItem;
 let type_subscription: vscode.Disposable | undefined;
 
 export function activate(context: vscode.ExtensionContext): void {
+    interface ModeConfig {
+        readonly name: string;
+        readonly capturing: boolean;
+    }
+    interface ModeConfigExtra extends ModeConfig {
+        readonly [key: string]: unknown
+    };
+
     const modalcode_settings = vscode.workspace.getConfiguration("modalcode");
 
-    const modes_properties = modalcode_settings.get("modes");
-    if (modes_properties === null) {
+    const modes_config = modalcode_settings.get("modes");
+    if (modes_config === null) {
         vscode.window.showInformationMessage("ModalCode: 'modalcode.modes' cannot be null");
         return;
     }
-    if (!Array.isArray(modes_properties)) {
-        vscode.window.showErrorMessage(`ModalCode: invalid 'modalcode.modes' setting type, expected 'array' but got '${typeof modes_properties}'`);
+    if (!Array.isArray(modes_config)) {
+        vscode.window.showErrorMessage(`ModalCode: invalid 'modalcode.modes' setting type, expected 'array' but got '${typeof modes_config}'`);
         return;
     }
-    if (modes_properties.length === 0) {
+    if (modes_config.length === 0) {
         vscode.window.showInformationMessage("ModalCode: no modes were defined");
         return;
     }
 
+    const MIN_NAME_LENGTH = 1;
     const MAX_NAME_LENGTH = 16;
 
     non_capturing_modes_start_index = 0;
-    for (let mode_properties_index = 0; mode_properties_index < modes_properties.length; mode_properties_index += 1) {
-        const mode_properties = modes_properties[mode_properties_index];
-        if (typeof mode_properties !== "object") {
-            vscode.window.showErrorMessage(`ModalCode: invalid mode properties type, expected 'object' but got '${typeof mode_properties}'`);
+    for (let config_index = 0; config_index < modes_config.length; config_index += 1) {
+        const mode_config = modes_config[config_index];
+
+        if (typeof mode_config !== "object") {
+            vscode.window.showErrorMessage(`ModalCode: invalid mode properties type, expected 'object' but got '${typeof mode_config}'`);
             return;
         }
-        if (mode_properties === null) {
+        if (mode_config === null) {
             vscode.window.showErrorMessage("ModalCode: mode properties cannot be null");
             return;
         }
 
-        if (!("name" in mode_properties)) {
+        if (!("name" in mode_config)) {
             vscode.window.showErrorMessage("ModalCode: missing 'name' mode property");
             return;
         }
-        if (mode_properties.name === null) {
+        if (mode_config.name === null) {
             vscode.window.showErrorMessage("ModalCode: 'name' mode property cannot be null");
             return;
         }
-        if (typeof mode_properties.name !== "string") {
-            vscode.window.showErrorMessage(`ModalCode: invalid 'name' mode property, expected 'string' but got '${typeof mode_properties.name}'`);
+        if (typeof mode_config.name !== "string") {
+            vscode.window.showErrorMessage(`ModalCode: invalid 'name' mode property, expected 'string' but got '${typeof mode_config.name}'`);
             return;
         }
-        if (mode_properties.name.length === 0) {
-            vscode.window.showErrorMessage("ModalCode: 'name' mode property cannot be empty");
+        if (mode_config.name.length < MIN_NAME_LENGTH) {
+            vscode.window.showErrorMessage(`ModalCode: 'name' mode property must be at least ${MIN_NAME_LENGTH} characters`);
             return;
         }
-        if (mode_properties.name.length > MAX_NAME_LENGTH) {
+        if (mode_config.name.length > MAX_NAME_LENGTH) {
             vscode.window.showErrorMessage(`ModalColde: 'name' mode property cannot be longer than ${MAX_NAME_LENGTH} characters`);
             return;
         }
 
-        if (!("capturing" in mode_properties)) {
+        if (!("capturing" in mode_config)) {
             vscode.window.showErrorMessage("ModalCode: missing 'capturing' mode property");
             return;
         }
-        if (mode_properties.capturing === null) {
+        if (mode_config.capturing === null) {
             vscode.window.showErrorMessage("ModalCode: 'capturing' mode property cannot be null");
             return;
         }
-        if (typeof mode_properties.capturing !== "boolean") {
-            vscode.window.showErrorMessage(`ModalCode: invalid 'capturing' mode property, expected 'boolean' but got '${typeof mode_properties.name}'`);
+        if (typeof mode_config.capturing !== "boolean") {
+            vscode.window.showErrorMessage(`ModalCode: invalid 'capturing' mode property, expected 'boolean' but got '${typeof mode_config.name}'`);
             return;
         }
 
-        type ModePropertiesExtra = ModeProperties & { [key: string]: unknown };
-        const { name, capturing, ...unexpected_properties } = mode_properties as ModePropertiesExtra;
+        const { name, capturing, ...unexpected_properties } = mode_config as ModeConfigExtra;
         if (Object.keys(unexpected_properties).length > 0) {
             for (const unexpected_property in unexpected_properties) {
                 vscode.window.showErrorMessage(`ModalCode: unexpected '${unexpected_property}' mode property for mode '${name}'`);
@@ -107,8 +117,8 @@ export function activate(context: vscode.ExtensionContext): void {
             return;
         }
 
-        for (let mode_index = 0; mode_index < mode_properties_index; mode_index += 1) {
-            const defined_mode_name = (modes_properties[mode_index] as ModeProperties).name;
+        for (let mode_index = 0; mode_index < config_index; mode_index += 1) {
+            const defined_mode_name = (modes_config[mode_index] as ModeConfig).name;
             if (defined_mode_name === name) {
                 vscode.window.showErrorMessage(`ModalCode: found duplicate mode '${defined_mode_name}'`);
                 return;
@@ -125,13 +135,13 @@ export function activate(context: vscode.ExtensionContext): void {
         return;
     }
 
-    modes = Array(modes_properties.length);
+    modes = Array(modes_config.length);
     let capturing_modes_start_index = 0;
     non_capturing_modes_start_index = modes.length;
 
     modes_names = Array(modes.length);
-    for (let mode_index = 0; mode_index < modes_properties.length; mode_index += 1) {
-        const { name, capturing } = modes_properties[mode_index] as ModeProperties;
+    for (let mode_index = 0; mode_index < modes_config.length; mode_index += 1) {
+        const { name, capturing } = modes_config[mode_index] as ModeConfig;
         const mode = new Mode(name);
         if (capturing) {
             modes[capturing_modes_start_index] = mode;
@@ -143,22 +153,22 @@ export function activate(context: vscode.ExtensionContext): void {
         modes_names[mode_index] = name;
     }
 
-    const select_mode_command = vscode.commands.registerCommand("modalcode.select_mode", select_mode);
-    const enter_mode_command = vscode.commands.registerCommand("modalcode.enter_mode", enter_mode);
+    const select_mode_command = vscode.commands.registerCommand(SELECT_MODE_COMMAND, select_mode);
+    const enter_mode_command = vscode.commands.registerCommand(ENTER_MODE_COMMAND, enter_mode);
 
     status_bar_item = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 9999999999);
-    status_bar_item.command = "modalcode.select_mode";
-    status_bar_item.tooltip = "Select mode";
+    status_bar_item.command = SELECT_MODE_COMMAND;
+    status_bar_item.tooltip = SELECT_MODE_TOOLTIP;
 
     const starting_mode = modes_names[0] as string;
-    enter_mode(starting_mode);
+    set_mode(starting_mode);
     status_bar_item.show();
 
     context.subscriptions.push(select_mode_command, enter_mode_command, status_bar_item);
 }
 
 export function deactivate(): void {
-    vscode.commands.executeCommand("setContext", "modalcode.mode", undefined);
+    vscode.commands.executeCommand("setContext", MODE_CONTEXT_KEY, undefined);
     if (type_subscription !== undefined) {
         type_subscription.dispose();
         type_subscription = undefined;
@@ -168,54 +178,71 @@ export function deactivate(): void {
 async function select_mode(): Promise<void> {
     const mode_name = await vscode.window.showQuickPick(modes_names, {
         canPickMany: false,
-        title: "Select mode",
-        placeHolder: "Select mode to enter",
+        title: SELECT_MODE_TOOLTIP,
+        placeHolder: SELECT_MODE_PLACEHOLDER,
     });
     if (mode_name === undefined) {
         return;
     }
 
-    enter_mode(mode_name);
+    set_mode(mode_name);
 }
 
-function enter_mode(mode_name: string): void {
-    let mode_index = 0;
-    for (; mode_index < non_capturing_modes_start_index; mode_index += 1) {
-        const mode = modes[mode_index] as Mode;
-        if (mode.name !== mode_name) {
-            continue;
-        }
+function enter_mode(name: unknown | undefined): void {
+    if (name === undefined) {
+        vscode.window.showErrorMessage("ModalCode: missing name argument");
+        return;
+    }
+    if (typeof name !== "string") {
+        vscode.window.showErrorMessage("ModalCode: name must be a string");
+        return;
+    }
 
-        if (type_subscription === undefined) {
-            try {
-                type_subscription = vscode.commands.registerCommand("type", () => {
-                    // disabling the 'type' command
-                });
-            } catch {
-                vscode.window.showErrorMessage(`ModalCode: cannot enter '${mode.name}' because the 'type' command is already registered`);
+    set_mode(name);
+}
+
+function set_mode(name: string): void {
+    let mode: Mode;
+    search_mode: {
+        let mode_index = 0;
+        for (; mode_index < non_capturing_modes_start_index; mode_index += 1) {
+            mode = modes[mode_index] as Mode;
+            if (mode.name !== name) {
+                continue;
             }
+
+            if (type_subscription === undefined) {
+                try {
+                    type_subscription = vscode.commands.registerCommand("type", () => {
+                        // disabling the 'type' command
+                    });
+                } catch {
+                    vscode.window.showErrorMessage(`ModalCode: cannot enter '${mode.name}' because the 'type' command is already registered`);
+                    return;
+                }
+            }
+
+            break search_mode;
         }
 
-        status_bar_item.text = mode.text;
-        vscode.commands.executeCommand("setContext", "modalcode.mode", mode.name);
+        for (; mode_index < modes.length; mode_index += 1) {
+            mode = modes[mode_index] as Mode;
+            if (mode.name !== name) {
+                continue;
+            }
+
+            if (type_subscription !== undefined) {
+                type_subscription.dispose();
+                type_subscription = undefined;
+            }
+
+            break search_mode;
+        }
+
+        vscode.window.showErrorMessage(`ModalCode: mode '${name}' not found`);
         return;
     }
 
-    for (; mode_index < modes.length; mode_index += 1) {
-        const mode = modes[mode_index] as Mode;
-        if (mode.name !== mode_name) {
-            continue;
-        }
-
-        if (type_subscription !== undefined) {
-            type_subscription.dispose();
-            type_subscription = undefined;
-        }
-
-        status_bar_item.text = mode.text;
-        vscode.commands.executeCommand("setContext", "modalcode.mode", mode.name);
-        return;
-    }
-
-    vscode.window.showErrorMessage(`ModalCode: mode '${mode_name}' not found`);
+    status_bar_item.text = mode.text;
+    vscode.commands.executeCommand("setContext", MODE_CONTEXT_KEY, mode.name);
 }